+++ src/runtime/events.rs
@@ -9,28 +9,32 @@ pub use crate::display::events::{
     CrossEvent, KeyCode, KeyEvent, KeyModifiers, MouseButton, MouseEvent, MouseEventKind,
 };
 
+/// An event raised by the app state.
 pub enum Event<T> {
+    /// A keyboard event.
     Key(KeyEvent),
+    /// A mouse event.
     Mouse(MouseEvent),
+    /// Resize event.
     Resize(Size),
+    /// User defined value was sent.
     User(T),
+    /// Replace the current widget tree in the [`AppState`].
     ReplaceWidgets(Vec<WidgetNode>),
+    /// Terminate the run loop in the app state.
     Quit,
 }
 
 impl<T> Event<T> {
+    /// Ctrl+c was pressed (useful to terminate the application).
     pub fn ctrl_c(&self) -> bool {
         matches!(self, Event::Key(KeyEvent { code: KeyCode::Char('c'), modifiers: KeyModifiers::CONTROL }))
     }
 
-    pub fn scroll_up(&self) -> bool {
-        matches!(self, Event::Mouse(MouseEvent { kind: MouseEventKind::ScrollUp, .. }))
-    }
-
-    pub fn scroll_down(&self) -> bool {
-        matches!(self, Event::Mouse(MouseEvent { kind: MouseEventKind::ScrollDown, .. }))
-    }
-
+    /// If the event was a mouse button down event, return
+    /// * `ScreenPos`
+    /// * `MouseButton`
+    /// * `KeyModifiers`
     pub fn mouse_down(&self) -> Option<(ScreenPos, MouseButton, KeyModifiers)> {
         match self {
             Event::Mouse(MouseEvent { kind: MouseEventKind::Down(btn), column, row, modifiers }) => {
@@ -40,6 +44,10 @@ impl<T> Event<T> {
         }
     }
 
+    /// If the event was a mouse button up event, return
+    /// * `ScreenPos`
+    /// * `MouseButton`
+    /// * `KeyModifiers`
     pub fn mouse_up(&self) -> Option<(ScreenPos, MouseButton, KeyModifiers)> {
         match self {
             Event::Mouse(MouseEvent { kind: MouseEventKind::Up(btn), column, row, modifiers }) => {
@@ -49,6 +57,7 @@ impl<T> Event<T> {
         }
     }
 
+    /// If the mouse was moved return `(ScreenPos, KeyModifiers)`.
     pub fn mouse_move(&self) -> Option<(ScreenPos, KeyModifiers)> {
         match self {
             Event::Mouse(MouseEvent { kind: MouseEventKind::Moved, column, row, modifiers }) => {
@@ -58,6 +67,12 @@ impl<T> Event<T> {
         }
     }
 
+    /// If a mouse button was pressed and the mouse moved then this qualifies as a mouse-drag
+    /// event.
+    /// This returns:
+    /// * `ScreenPos`
+    /// * `MouseButton`
+    /// * `KeyModifiers`
     pub fn mouse_drag(&self) -> Option<(ScreenPos, MouseButton, KeyModifiers)> {
         match self {
             Event::Mouse(MouseEvent { kind: MouseEventKind::Drag(btn), column, row, modifiers }) => {
@@ -67,7 +82,8 @@ impl<T> Event<T> {
         }
     }
 
-    pub fn mouse_scroll_up(&self) -> Option<(ScreenPos, KeyModifiers)> {
+    /// Mouse scroll up, returns `(ScreenPos, KeyModifiers)`.
+    pub fn scroll_up(&self) -> Option<(ScreenPos, KeyModifiers)> {
         match self {
             Event::Mouse(MouseEvent { kind: MouseEventKind::ScrollUp, column, row, modifiers }) => {
                 Some((ScreenPos::new(*column, *row), *modifiers))
@@ -76,7 +92,8 @@ impl<T> Event<T> {
         }
     }
 
-    pub fn mouse_scroll_down(&self) -> Option<(ScreenPos, KeyModifiers)> {
+    /// Mouse scroll down, returns `(ScreenPos, KeyModifiers)`.
+    pub fn scroll_down(&self) -> Option<(ScreenPos, KeyModifiers)> {
         match self {
             Event::Mouse(MouseEvent { kind: MouseEventKind::ScrollDown, column, row, modifiers }) => {
                 Some((ScreenPos::new(*column, *row), *modifiers))
@@ -85,6 +102,16 @@ impl<T> Event<T> {
         }
     }
 
+    /// If a keyboard character was pressed return the keycode.
+    pub fn get_keycode(&self) -> Option<KeyCode> {
+        if let Event::Key(KeyEvent { code, .. }) = self {
+            Some(*code)
+        } else {
+            None
+        }
+    }
+
+    /// If the expected character was pressed return true.
     pub fn is_char(&self, expected: char) -> bool {
         if let Event::Key(KeyEvent { code: KeyCode::Char(c), .. }) = self {
             expected.eq(c)
@@ -93,6 +120,7 @@ impl<T> Event<T> {
         }
     }
 
+    /// A user defined value was received.
     pub fn user(self) -> Option<T> {
         match self {
             Event::User(val) => Some(val),