+++ CHANGES.md
@@ -1,3 +1,6 @@
+* 0.2.0
+    * `runtime::Event::scroll_up` and `runtime::Event::scroll_down` now returns
+      the screen position and any key modifiers
 * 0.1.2
     * `WidgetContainer::by_id` is made more ergonomic and can now called with a
       string slice now.